implemented user cart info fetch and all cart info fetch for admin

+++ src/main/java/com/springproject/profEcomWebApp/security/response/UserInfoResponse.java
@@ -4,30 +4,30 @@ import java.util.List;
 
 public class UserInfoResponse {
 
-    private Integer id;
+    private Long id;
     private String jwtToken;
 
     private String username;
     private List<String> roles;
 
-    public UserInfoResponse(Integer id, String username, List<String> roles, String jwtToken) {
+    public UserInfoResponse(Long id, String username, List<String> roles, String jwtToken) {
         this.id = id;
         this.username = username;
         this.roles = roles;
         this.jwtToken = jwtToken;
     }
 
-    public UserInfoResponse(Integer id, String username, List<String> roles) {
+    public UserInfoResponse(Long id, String username, List<String> roles) {
         this.id = id;
         this.username = username;
         this.roles = roles;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
